requirements.txt
flask
google-cloud-speech
numpy
nltk
spacy
